move samplers and collaters to separate submodules
