+++ src/deep_taxon/nn/loader.py
@@ -12,12 +12,15 @@ import pytorch_lightning as pl
 import torch.nn.functional as F
 import torch
 import numpy as np
-from torch.utils.data import DataLoader, Dataset, Sampler, Subset, SequentialSampler
+from torch.utils.data import DataLoader, Dataset, Subset
 from sklearn.model_selection import train_test_split
 from sklearn.utils import check_random_state
 
+from .collaters import get_collater
+from .samplers import DSSampler, WORSampler
 from ..sequence import LazyWindowChunkedDIFile, DeepIndexFile, chunk_sequence, lazy_chunk_sequence
 from ..utils import parse_seed, distsplit, balsplit, get_logger, log
+
 
 def print_mem(msg=None):
     pid = os.getpid()
@@ -306,305 +309,6 @@ def check_loaded_sequences(argv=None):
     logger.info(f"{correct} ({rc} revcomped) samples of {n_samples} existed in the respective genomes returned by the loader")
 
 
-class SplitCollater:
-
-    def __init__(self, padval, freq=1.0, factors=[2, 4, 8]):
-        self.padval = padval
-        self.freq = freq
-        self.factors = factors
-
-    def __call__(self, samples):
-        maxlen = 0
-        l_idx = -1
-        if isinstance(samples, tuple):
-            samples = [samples]
-        for i, X, y, seq_id in samples:
-            if maxlen < X.shape[l_idx]:
-                maxlen = X.shape[l_idx]
-        X_ret = list()
-        y_ret = list()
-        idx_ret = list()
-        size_ret = list()
-        seq_id_ret = list()
-        for i, X, y, seq_id in samples:
-            dif = maxlen - X.shape[l_idx]
-            X_ = X
-            if dif > 0:
-                X_ = F.pad(X, (0, dif), value=self.padval)
-            X_ret.append(X_)
-            y_ret.append(y)
-            size_ret.append(X.shape[l_idx])
-            idx_ret.append(i)
-            seq_id_ret.append(seq_id)
-        X_ret = torch.stack(X_ret)
-        y_ret = torch.stack(y_ret)
-        size_ret = torch.tensor(size_ret)
-        idx_ret = torch.tensor(idx_ret)
-        seq_id_ret = torch.tensor(seq_id_ret)
-
-        if self.freq == 1.0 or rs.rand() < self.freq:
-            f = factors[rs.randint(len(factors))]
-            y_ret = y_ret.repeat_interleave(f)
-            seq_id_ret = seq_id_ret.repeat_interleave(f)
-            idx_ret = idx_ret.repeat_interleave(f)
-            X_ret = X_ret.reshape((X_ret.shape[0] * f, X_ret.shape[0] // f))
-
-            q = lens // f
-            r = lens // f
-            n_bad_chunks = b.shape[-1]//f - q
-            bad_chunk_pos = torch.where(n_bad_chunks > 0)[0]
-            start_bad_chunks = bad_chunk_pos + q[bad_chunk_pos]
-
-        return (idx_ret, X_ret, y_ret, size_ret, seq_id_ret)
-
-
-class SeqCollater:
-
-    def __init__(self, padval):
-        self.padval = padval
-
-    def __call__(self, samples):
-        maxlen = 0
-        l_idx = -1
-        if isinstance(samples, tuple):
-            samples = [samples]
-        for i, X, y, seq_id in samples:
-            if maxlen < X.shape[l_idx]:
-                maxlen = X.shape[l_idx]
-        X_ret = list()
-        y_ret = list()
-        idx_ret = list()
-        size_ret = list()
-        seq_id_ret = list()
-        for i, X, y, seq_id in samples:
-            dif = maxlen - X.shape[l_idx]
-            X_ = X
-            if dif > 0:
-                X_ = F.pad(X, (0, dif), value=self.padval)
-            X_ret.append(X_)
-            y_ret.append(y)
-            size_ret.append(X.shape[l_idx])
-            idx_ret.append(int(i))
-            seq_id_ret.append(int(seq_id))
-        X_ret = torch.stack(X_ret)
-        y_ret = torch.stack(y_ret)
-        size_ret = torch.tensor(size_ret)
-        idx_ret = torch.tensor(idx_ret)
-        seq_id_ret = torch.tensor(seq_id_ret)
-        return (idx_ret, X_ret, y_ret, size_ret, seq_id_ret)
-
-
-class TrainingSeqCollater:
-
-    def __init__(self, padval, seq_dtype=torch.uint8):
-        self.padval = padval
-        self.seq_dtype = seq_dtype
-
-    def __call__(self, samples):
-        maxlen = 0
-        l_idx = -1
-        if isinstance(samples, tuple):
-            samples = [samples]
-        for i, X, y, seq_id in samples:
-            if maxlen < X.shape[l_idx]:
-                maxlen = X.shape[l_idx]
-        X_ret = list()
-        y_ret = list()
-        for i, X, y, seq_id in samples:
-            dif = maxlen - X.shape[l_idx]
-            X_ = X
-            if dif > 0:
-                X_ = F.pad(X, (0, dif), value=self.padval)
-            X_ret.append(X_)
-            y_ret.append(y)
-        X_ret = torch.stack(X_ret, out=torch.zeros(len(X_ret), len(X_ret[0]), dtype=self.seq_dtype))
-        y_ret = torch.stack(y_ret, out=torch.zeros(len(X_ret), dtype=torch.int32))
-        return (X_ret, y_ret)
-
-
-def _check_collater(padval, seq_collater):
-    if seq_collater is not None:
-        return seq_collater
-    elif padval is not None:
-        return SeqCollater(padval)
-    else:
-        raise ValueError("must specify padval or seq_collater")
-
-class GraphCollater:
-
-    def __init__(self, node_ids, padval=None, seq_collater=None):
-        self.collater = _check_collater(padval, seq_collater)
-        self.node_ids = torch.as_tensor(node_ids, dtype=torch.long)
-
-    def __call__(self, samples):
-        """
-        A function to collate samples and return a sub-distance matrix
-        """
-        idx_ret, X_ret, y_idx, size_ret, seq_id_ret = self.collater(samples)
-
-        # Get distances
-        y_ret = self.node_ids[y_idx]
-        return (idx_ret, X_ret, y_ret, size_ret, seq_id_ret)
-
-
-class DistanceCollater:
-
-    def __init__(self, dmat, padval=None, seq_collater=None):
-        self.collater = _check_collater(padval, seq_collater)
-        if len(dmat.shape) == 1:
-            from scipy.spatial.distance import squareform
-            dmat = squareform(dmat)
-        #self.dmat = torch.as_tensor(dmat, dtype=torch.float).pow(2)
-        #self.dmat = torch.as_tensor(dmat, dtype=torch.float).sqrt()
-        self.dmat = dmat
-        if not isinstance(self.dmat, torch.Tensor):
-            self.dmat = torch.from_numpy(self.dmat)
-
-        self.dmat /= self.dmat.max()
-
-
-    def __call__(self, samples):
-        """
-        A function to collate samples and return a sub-distance matrix
-        """
-        X_ret, y_idx = self.collater(samples)
-
-        # Get distances
-        y_idx = y_idx.long()
-        y_ret = self.dmat[y_idx][:, y_idx]
-        return X_ret, y_ret
-
-
-class TnfCollater:
-    def __init__(self, vocab):
-        self.bases = 4**torch.arange(4)
-        rcmap = torch.tensor([3, 2, 1, 0])
-        canonical = list()
-        noncanonical = list()
-        palindromes = list()
-        seen = torch.zeros(256, dtype=bool)
-        for i in range(256):
-            if seen[i]:
-                continue
-            ar = torch.zeros(4, dtype=int)
-            ar[3], r = divmod(i, 64)
-            ar[2], r = divmod(r, 16)
-            ar[1], ar[0] = divmod(r, 4)
-            rc = rcmap[ar.flip(0)]
-            rc_i = rc.matmul(self.bases)
-            if i < rc_i:
-                canonical.append(i)
-                noncanonical.append(rc_i)
-            elif rc_i < i:
-                canonical.append(rc_i)
-                noncanonical.append(i)
-            else:
-                palindromes.append(i)
-            seen[i] = True
-            seen[rc_i] = True
-        self.canonical = torch.tensor(canonical)
-        self.noncanonical = torch.tensor(noncanonical)
-        self.palindromes = torch.tensor(palindromes)
-
-        # calculate a map to convert DNA characters into 0-4 encoding
-        self.cmap = torch.zeros(128, dtype=int) - 1
-        count = 0
-        self.padval = None
-        for i, c in enumerate(vocab):
-            if c == 'A':
-                self.cmap[i] = 0            # A
-                count += 1
-            elif c == 'T':
-                self.cmap[i] = 3            # T
-                count += 1
-            elif c == 'C':
-                self.cmap[i] = 1            # C
-                count += 1
-            elif c == 'G':
-                self.cmap[i] = 2            # G
-                count += 1
-            elif c == 'N':
-                self.padval = i
-                count += 1
-            if count == 5:
-                break
-        if self.padval is None:
-            raise ValueError("Could not find 'N' character in vocab -- this is needed to pad sequences")
-
-
-    def __call__(self, samples):
-        l_idx = -1
-        if isinstance(samples, tuple):
-            samples = [samples]
-
-        maxlen = 0
-        for i, X, y, seq_id in samples:
-            if maxlen < X.shape[l_idx]:
-                maxlen = X.shape[l_idx]
-
-        X_ret = list()
-        y_ret = list()
-        idx_ret = list()
-        size_ret = list()
-        seq_id_ret = list()
-        for i, X, y, seq_id in samples:
-            dif = maxlen - X.shape[l_idx]
-            X_ = X
-            if dif > 0:
-                X_ = F.pad(X, (0, dif), value=self.padval)
-            X_ret.append(X_)
-            y_ret.append(y)
-            size_ret.append(X.shape[l_idx])
-            idx_ret.append(i)
-            seq_id_ret.append(seq_id)
-
-        # calculate tetranucleotide frequency
-        chunks = torch.stack(X_ret)
-
-        ## 1. hash 4-mers
-        __seq = self.cmap[chunks]
-        i4mers = torch.stack([__seq[:, 0:-3], __seq[:, 1:-2], __seq[:, 2:-1], __seq[:, 3:]], axis=2)
-        mask = torch.any(i4mers < 0, axis=2)
-        h4mers = i4mers.matmul(self.bases)       # hashed 4-mers
-        h4mers[mask] = 256    # use 257 to mark any 4-mers that had ambiguous nucleotides
-
-        ## 2. count hashed 4-mers i.e. count integers from between 0-257 inclusive
-        tnf = torch.zeros((32, 257), dtype=float)
-        for i in range(tnf.shape[0]):
-            counts = torch.bincount(h4mers[i], minlength=257)
-            tnf[i] = counts/i4mers.shape[1]
-
-        ## 3. merge canonical 4-mers
-        canon_tnf = torch.zeros((32, 136))
-        canon_tnf[:, :len(self.canonical)] = tnf[:, self.canonical] + tnf[:, self.noncanonical]
-        canon_tnf[:, len(self.canonical):] = tnf[:, self.palindromes]
-
-        X_ret = canon_tnf
-        y_ret = torch.stack(y_ret)
-        size_ret = torch.tensor(size_ret)
-        idx_ret = torch.tensor(idx_ret)
-        seq_id_ret = torch.tensor(seq_id_ret)
-
-        return (idx_ret, X_ret, y_ret, size_ret, seq_id_ret)
-
-
-def get_collater(dataset, inference=False, condensed=False):
-    if dataset.tnf:
-        return TnfCollater(dataset.vocab)
-    elif dataset.manifold:
-        if condensed:
-            return TrainingSeqCollater(dataset.padval)
-        else:
-            return DistanceCollater(dataset.difile.distances, seq_collater=TrainingSeqCollater(dataset.padval))
-    elif dataset.graph:
-        return GraphCollater(dataset.difile.node_ids, seq_collater=TrainingSeqCollater(dataset.padval))
-    else:
-        if inference:
-            return SeqCollater(dataset.padval)
-        else:
-            return TrainingSeqCollater(dataset.padval)
-
-
 def get_loader(dataset, inference=False, **kwargs):
     """
     Return a DataLoader that loads data from the given Dataset
@@ -628,140 +332,6 @@ def get_loader(dataset, inference=False, **kwargs):
     return DataLoader(orig_dataset, collate_fn=collater, **kwargs)
 
 
-def check_rng(rng):
-    if rng is None:
-        return np.random.default_rng()
-    elif isinstance(rng, (int, np.integer)):
-        return np.random.default_rng(rng)
-    return rng
-
-
-class WORHelper:
-
-    def __init__(self, length, rng=None):
-        self.period = length
-        self.rng = check_rng(rng)
-        self.indices = np.arange(length)
-        self.curr_max = length - 1
-
-    def sample(self):
-        idx = self.rng.integers(self.curr_max + 1)
-        ret = self.indices[idx]
-        end = self.curr_max
-        self.indices[end], self.indices[idx] = self.indices[idx], self.indices[end]
-        self.curr_max = (self.curr_max - 1) % self.period
-        return ret
-
-
-class WORSampler(Sampler):
-    """Without Replacement Sampler"""
-
-    def __init__(self, length, rng=None, n_partitions=1, part_smplr_rng=None, max_samples=None):
-        """
-        Args:
-            n_partitions :          number of deterministic partitions to break up dataset into
-            part_smplr_rng :        Partition Sampler RNG
-        """
-        super().__init__(None)
-        self.rng = check_rng(rng)
-        dtype = np.uint32
-        if length > (2**32 - 1):
-            dtype = np.uint64
-
-
-        # trim will clip extra samples (i.e. length % size) so that each
-        # rank has the same number of samples.
-        # Use this later if we decide we don't want to trim tail.
-        self.indices = np.arange(length, dtype=dtype)
-
-        self.part_sampler = WORHelper(n_partitions, rng=part_smplr_rng)
-
-        # set an initial value for curr_part (i.e. current partition)
-        # so we can calculate lenght in __len__
-        self.curr_part = self.part_sampler.sample()
-
-        self.__len = (len(self.indices) - self.curr_part - 1) // self.part_sampler.period + 1
-        if max_samples is not None:
-            self.__len = min(self.__len, max_samples)
-        self.max_samples = max_samples
-        self.i = self.__len
-
-        self.curr_len = 0
-
-    def __iter__(self):
-        self.curr_len = (len(self.indices) - self.curr_part - 1) // self.part_sampler.period + 1 # len(self)
-        if self.max_samples is not None:
-            self.i = 0
-        return self
-
-    def __len__(self):
-        return self.__len
-
-    def __next__(self):
-        if self.max_samples is not None:
-            if self.i < self.__len:
-                self.i += 1
-            else:
-                raise StopIteration
-
-        if self.curr_len == 0:
-            self.curr_part = self.part_sampler.sample()
-            raise StopIteration
-
-        idx = self.rng.integers(self.curr_len) * self.part_sampler.period + self.curr_part
-        end = (self.curr_len - 1) * self.part_sampler.period + self.curr_part
-        ret = self.indices[idx]
-        self.indices[end], self.indices[idx] = self.indices[idx], self.indices[end]
-        self.curr_len -= 1
-        return ret
-
-
-class DSSampler(SequentialSampler):
-    """Distributed Sequential Sampler"""
-
-    def __init__(self, length, n_partitions=1, part_smplr_rng=None, max_samples=None):
-        """
-        Args:
-            n_partitions :          number of deterministic partitions to break up dataset into
-            part_smplr_rng :        Partition Sampler RNG
-        """
-        super().__init__(None)
-        self.part_sampler = WORHelper(n_partitions, rng=part_smplr_rng)
-
-        # set an initial value for curr_part (i.e. current partition)
-        # so we can calculate lenght in __len__
-        self.curr_part = self.part_sampler.sample()
-        self.start = 0
-        self.end = length
-        self.__len = (self.end - self.start - self.curr_part - 1) // self.part_sampler.period + 1
-        if max_samples is not None:
-            self.__len = min(self.__len, max_samples)
-        self.max_samples = max_samples
-        self.i = self.__len
-
-
-    def __iter__(self):
-        self.__it = iter(range(self.start + self.curr_part, self.end, self.part_sampler.period))
-        if self.max_samples is not None:
-            self.i = 0
-        return self
-
-    def __len__(self):
-        return self.__len
-
-    def __next__(self):
-        try:
-            if self.max_samples is not None:
-                if self.i < self.__len:
-                    self.i += 1
-                else:
-                    raise StopIteration
-            return next(self.__it)
-        except StopIteration:
-            self.curr_part = self.part_sampler.sample()
-            raise StopIteration
-
-
 class SubsetDataLoader(DataLoader):
 
     def __init__(self, dataset, train=False, validate=False, test=False, **kwargs):
@@ -779,9 +349,6 @@ class SubsetDataLoader(DataLoader):
         self.dataset.set_subset(train=self.train, validate=self.validate, test=self.test)
         return super().__iter__()
 
-
-from torch.utils.data import Dataset, DataLoader
-import torch
 
 class fast_dataset(Dataset):
     def __init__(self, num_samples, sample_length):
